requirements.txt
streamlit
groq
requests
python-dotenv
sentence-transformers
faiss-cpu
numpy
pymupdf